microservices/lambdas/pimbda/requirements.txt
newspaper3k==0.2.8
botocore==1.7.39
certifi==2017.7.27.1
chalice==1.6.1
click==6.6
cycler==0.10.0
decorator==4.1.2
docutils==0.14
jmespath==0.9.3
olefile==0.44
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyWavelets==0.5.2
six==1.11.0
typing==3.5.3.0
json-tricks==3.8.0
requests==2.18.4
lxml==3.8.0
pymongo==3.7.2
dnspython==1.15.0